+++ vite.config.ts
@@ -27,6 +27,6 @@ export default defineConfig({
   test: {
     environment: 'node',
     setupFiles: ['./test/setup.ts'],
-    include: ['test/**/*.test.ts'],
+    include: ['test/**/*.{test,spec}.{ts,tsx}'],
   },
 });